simple_webbrowser: my (licensed) gist!
Gist at: https://gist.github.com/MF366-Coding/a91f1ea111fc22bf871afc4038f27166

More info: My gist allows for a more simple use of the webbrowser module, allowing the developer to use functions such as Google, Bing, Ecosia, that goes directly to that search engine and searches your query automatically. WriterClassic.py is basicaly the same but now uses that gist of mine as a class, which makes it a bit more clear. Also, since that gist is licensed, I placed that license inside the Python file for WriterClassic. That is why the file became so large, when it comes to the amount of lines it has. It also contains the gist as a class but I already mencioned that.

There's no reason to place this as a release since no features were visually changed, added or removed.

See you soon, with a new release! :wave:

Thanks for using WriterClassic and for reading all of this! :heart: